backend/requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
itypes==1.2.0
Markdown==3.7
openapi-codec==1.3.2
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.0
pytz==2024.2
setuptools==75.6.0
simplejson==3.19.3
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
zipp==3.20.2
